#101: Move granular utils to their own classes.
